Expose inherited part of action environment to Starlark tests
Starlark tests can use the new `Action#env_inherit` field to get a list
of environment variables the action inherits from the client
environment. This list is ensured to be accurate by being derived
directly from `getEffectiveEnvironment`.

Since this requires passing a fake client environment into
`getEffectiveEnvironment`, the method is refactored to take in a getter
function (e.g. `Map::get`) rather than an actual `Map`. As a side
effect, this ensures that actions cannot iterate of the client
environment, but are limited to query it for specific variables.