Add a native image of turbine to the prebuilt Java tools
A simple local experiment shows that the time for all Java header
compilation actions required for `//src:bazel-dev` decreases by a factor
of ~4.5 when using a native image of turbine instead of jar.

As a first step towards using the native image in Java toolchains, ship
it as part of the prebuilt Java tools by using rules_graalvm.